chore: add comments on pr suggesting formatting
This commit adds a step to the lua-format-check-pr job that will leave a
comment with "suggested changes" when formatting needs to be applied.
This can then be applied with one click (per file).

Alternatively, I could change it to automattically add a commit that
formats the code. Several PRs have gotten through without proper
formatting, which makes all future PRs fail the checks.